// frontend/src/core/api/generated/model/productionItemViewStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */

export type ProductionItemViewStatus =
  (typeof ProductionItemViewStatus)[keyof typeof ProductionItemViewStatus]

export const ProductionItemViewStatus = {
  IN_PROGRESS: 'IN_PROGRESS',
  NEARLY_DONE: 'NEARLY_DONE',
  DELAYED: 'DELAYED',
  COMPLETED: 'COMPLETED',
} as const